Test should not fail because of exception raised
Since commit 4b36d0c08fc55d45ca4025daf4bad5c8349efcb4 in net-ssh, this
test fails. Since an exception is raised in the middle of the story, the
last steps are not processed, and assert_scripted fails.

Given an exception is raised in the middle of exception, it is expected
that the story cannot finish completely.