Handle remotes closes without eof
Some ssh-server (APC Network Management Card AP9640) close the channel
without sending an eof. This is correctly handled by openssh, so NET:SCP
should also handle it.
Added a test for this case.
The changed/added code passes Rubocop with default settings, with the
exception of the length of the test name, wich ich intended to be long.